Display the "first" topic as the index page
As in, "Hello HTTP & HTML".  "First" is sorta random because it just
picks the first from the map.  Because we only have one topic for now
this is ok, but we'll have to change it later.  (In the "real" version
we'll need to pick the most recent or "currently active" topic, not just
any from the db.  We could maybe also have a "flag" in the db for that,
but I'm not quite sure about that.)